chore: convert build-zip.js to ts
