+++ tsconfig.json
@@ -7,6 +7,8 @@
         "strict": true,
         "rootDirs": ["src", "test"],
         "outDir": "dist/js",
-        "noEmitOnError": true
+        "noEmitOnError": true,
+        "esModuleInterop": true,
+        "resolveJsonModule": true
     }
 }
